requirements.txt
splat64[mips]==0.22.1
mapfile-parser>=2.3.5,<3.0.0
colorama
libyaz0

ipl3checksum>=1.2.0,<2.0.0
spimdisasm>=1.16.3,<2.0.0
